#79 [mod] 메뉴 추천뷰 폴더 위치 이동
