+++ client/app/src/main/java/com/kaer/menuw/presentation/home/menurecipe/RecipePageActivity.kt
@@ -6,7 +6,7 @@ import androidx.activity.viewModels
 import androidx.viewpager2.widget.ViewPager2
 import com.kaer.menuw.R
 import com.kaer.menuw.databinding.ActivityRecipePageBinding
-import com.kaer.menuw.presentation.home.refrigerator.recommend.MenuListViewModel.Companion.CHOOSE_MENU
+import com.kaer.menuw.presentation.recommend.MenuListViewModel.Companion.CHOOSE_MENU
 import com.kaer.menuw.util.base.BaseActivity
 import com.kaer.menuw.util.base.BaseDialog
 import dagger.hilt.android.AndroidEntryPoint
