+++ client/app/src/main/java/com/kaer/menuw/presentation/home/refrigerator/RefrigeratorFragment.kt
@@ -9,7 +9,7 @@ import com.kaer.menuw.databinding.FragmentRefrigeratorBinding
 import com.kaer.menuw.presentation.home.refrigerator.add.AddIngredientActivity
 import com.kaer.menuw.presentation.home.refrigerator.add.AddIngredientViewModel
 import com.kaer.menuw.presentation.home.refrigerator.add.SharedPreferenceManager
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.MenuCategoryActivity
+import com.kaer.menuw.presentation.recommend.category.MenuCategoryActivity
 import com.kaer.menuw.util.base.BaseDialog
 import com.kaer.menuw.util.base.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
