+++ client/app/src/main/java/com/kaer/menuw/presentation/recommend/MenuNotRecommendFragment.kt
@@ -1,4 +1,4 @@
-package com.kaer.menuw.presentation.home.refrigerator.recommend
+package com.kaer.menuw.presentation.recommend
 
 import android.content.Intent
 import android.os.Bundle
