+++ client/app/src/main/java/com/kaer/menuw/presentation/recommend/category/MenuCategoryActivity.kt
@@ -1,4 +1,4 @@
-package com.kaer.menuw.presentation.home.refrigerator.recommend.category
+package com.kaer.menuw.presentation.recommend.category
 
 import android.content.Intent
 import android.os.Bundle
@@ -9,12 +9,12 @@ import androidx.fragment.app.replace
 import com.kaer.menuw.R
 import com.kaer.menuw.databinding.ActivityMenuCategoryBinding
 import com.kaer.menuw.presentation.home.refrigerator.RefrigeratorFragment.Companion.INGREDIENT_ID_LIST
-import com.kaer.menuw.presentation.home.refrigerator.recommend.MenuListActivity
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.MenuCategoryViewModel.Companion.RECOMMEND_PAGE
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.MenuCategoryViewModel.Companion.TYPE_PAGE
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.model.RecommendRequestIntent
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.recipe.CategoryRecipeFragment
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.type.CategoryTypeFragment
+import com.kaer.menuw.presentation.recommend.MenuListActivity
+import com.kaer.menuw.presentation.recommend.category.MenuCategoryViewModel.Companion.RECOMMEND_PAGE
+import com.kaer.menuw.presentation.recommend.category.MenuCategoryViewModel.Companion.TYPE_PAGE
+import com.kaer.menuw.presentation.recommend.category.model.RecommendRequestIntent
+import com.kaer.menuw.presentation.recommend.category.recipe.CategoryRecipeFragment
+import com.kaer.menuw.presentation.recommend.category.type.CategoryTypeFragment
 import com.kaer.menuw.util.base.BaseActivity
 import dagger.hilt.android.AndroidEntryPoint
 import timber.log.Timber
