+++ client/app/src/main/java/com/kaer/menuw/presentation/recommend/category/recipe/CategoryRecipeAdapter.kt
@@ -1,17 +1,18 @@
-package com.kaer.menuw.presentation.home.refrigerator.recommend.category.type
+package com.kaer.menuw.presentation.recommend.category.recipe
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.kaer.menuw.R
-import com.kaer.menuw.databinding.ItemCategoryTypeBinding
+import com.kaer.menuw.databinding.ItemCategoryRecipeBinding
+import com.kaer.menuw.domain.entity.CategoryRecipe
 import com.kaer.menuw.util.ItemDiffCallback
 
-class CategoryTypeAdapter :
-    ListAdapter<String, CategoryTypeAdapter.CategoryTypeViewHolder>(
-        ItemDiffCallback<String>(
-            onItemsTheSame = { old, new -> old == new },
+class CategoryRecipeAdapter :
+    ListAdapter<CategoryRecipe, CategoryRecipeAdapter.CategoryRecipeViewHolder>(
+        ItemDiffCallback<CategoryRecipe>(
+            onItemsTheSame = { old, new -> old.recipeName == new.recipeName },
             onContentsTheSame = { old, new -> old == new }
         )
     ) {
@@ -20,18 +21,23 @@ class CategoryTypeAdapter :
     private var selectedPosition: Int = 10
 
     interface OnItemClickListener {
-        fun onItemClick(item: String, position: Int)
+        fun onItemClick(item: CategoryRecipe, position: Int)
     }
 
     fun setOnItemClickListener(listener: OnItemClickListener) {
         this.onItemClickListener = listener
     }
 
-    inner class CategoryTypeViewHolder(
-        private val binding: ItemCategoryTypeBinding
+    inner class CategoryRecipeViewHolder(
+        private val binding: ItemCategoryRecipeBinding
     ) : RecyclerView.ViewHolder(binding.root) {
-        fun onBind(data: String) {
-            binding.tvItemCategoryType.text = data
+        fun onBind(data: CategoryRecipe) {
+            binding.item = data
+            binding.ivItemRecipeImg.setImageResource(data.recipeImg)
+//             binding.root.setOnClickListener {
+//                 recipeSelection(binding, data.recipeName)
+//                 onItemClickListener?.let { it(data.recipeName) }
+//             }
 
             if (selectedPosition == absoluteAdapterPosition) {
                 changeItemColor(binding, true)
@@ -57,30 +63,30 @@ class CategoryTypeAdapter :
     }
 
     private fun changeItemColor(
-        binding: ItemCategoryTypeBinding,
+        binding: ItemCategoryRecipeBinding,
         selected: Boolean
     ) {
         when (selected) {
             true -> {
-                binding.layoutItemCategoryType.setBackgroundResource(R.drawable.shape_bright_fill_3_green_line_rect)
+                binding.layoutItemCategory.setBackgroundResource(R.drawable.shape_bright_fill_5_green_line_rect)
             }
 
             false -> {
-                binding.layoutItemCategoryType.setBackgroundResource(R.drawable.shape_green2_bright_fill_3_rect)
+                binding.layoutItemCategory.setBackgroundResource(R.drawable.shape_green2_bright_fill_5_rect)
             }
         }
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CategoryTypeViewHolder {
-        val binding = ItemCategoryTypeBinding.inflate(
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CategoryRecipeViewHolder {
+        val binding = ItemCategoryRecipeBinding.inflate(
             LayoutInflater.from(parent.context),
             parent,
             false
         )
-        return CategoryTypeViewHolder(binding)
+        return CategoryRecipeViewHolder(binding)
     }
 
-    override fun onBindViewHolder(holder: CategoryTypeViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: CategoryRecipeViewHolder, position: Int) {
         holder.onBind(currentList[position])
     }
 }
