+++ client/app/src/main/java/com/kaer/menuw/presentation/recommend/category/recipe/CategoryRecipeFragment.kt
@@ -1,4 +1,4 @@
-package com.kaer.menuw.presentation.home.refrigerator.recommend.category.recipe
+package com.kaer.menuw.presentation.recommend.category.recipe
 
 import android.os.Bundle
 import android.view.View
@@ -7,8 +7,8 @@ import androidx.recyclerview.widget.GridLayoutManager
 import com.kaer.menuw.R
 import com.kaer.menuw.databinding.FragmentCategoryRecipeBinding
 import com.kaer.menuw.domain.entity.CategoryRecipe
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.MenuCategoryViewModel
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.MenuCategoryViewModel.Companion.TYPE_PAGE
+import com.kaer.menuw.presentation.recommend.category.MenuCategoryViewModel
+import com.kaer.menuw.presentation.recommend.category.MenuCategoryViewModel.Companion.TYPE_PAGE
 import com.kaer.menuw.util.base.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
