+++ client/app/src/main/java/com/kaer/menuw/presentation/recommend/category/type/CategoryTypeFragment.kt
@@ -1,4 +1,4 @@
-package com.kaer.menuw.presentation.home.refrigerator.recommend.category.type
+package com.kaer.menuw.presentation.recommend.category.type
 
 import android.os.Bundle
 import android.view.View
@@ -6,8 +6,8 @@ import androidx.fragment.app.activityViewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.kaer.menuw.R
 import com.kaer.menuw.databinding.FragmentCategoryTypeBinding
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.MenuCategoryViewModel
-import com.kaer.menuw.presentation.home.refrigerator.recommend.category.MenuCategoryViewModel.Companion.RECOMMEND_PAGE
+import com.kaer.menuw.presentation.recommend.category.MenuCategoryViewModel
+import com.kaer.menuw.presentation.recommend.category.MenuCategoryViewModel.Companion.RECOMMEND_PAGE
 import com.kaer.menuw.util.base.BaseFragment
 import dagger.hilt.android.AndroidEntryPoint
